**后台管理系统** 功能完善——删除班级或者部门时，如果下面有学生或者员工，不允许删除并给前端传递错误信息

+++ web-ai-project02/tlias-web-management/src/main/java/com/itheima/service/impl/ClazzServiceImpl.java
@@ -2,6 +2,7 @@ package com.itheima.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.itheima.exception.ClassNotEmptyException;
 import com.itheima.mapper.ClazzMapper;
 import com.itheima.pojo.Clazz;
 import com.itheima.pojo.ClazzQueryParam;
@@ -42,6 +43,9 @@ public class ClazzServiceImpl implements ClazzService {
     @Override
     public void deleteById(Integer id) {
         // 如果班级下有学生，提示错误信息："对不起, 该班级下有学生, 不能直接删除"。
+        if (clazzMapper.getStudentCountByClazzId(id) > 0) {
+            throw new ClassNotEmptyException("对不起, 该班级下有学生, 不能直接删除");
+        }
         clazzMapper.deleteById(id);
     }
